ai-platform/services/retrieval_service/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
pydantic==2.7.4
qdrant-client==1.9.1
google-generativeai==0.5.4
